Updated logic branch for handleChessboardClickAction() method
The logic has been changed so that the "second-click" logic first
checks if the user has clicked the same chess piece that they
initially clicked in their "first-click" action. If so, just
un-highlight the chess piece and assume it is now their "first-click"
action again; otherwise check to see if the move is a Valid chess move.

The logic was flipped around, but now the least-processor-intensive
(checking if the first and second-click pieces are the same) logic
is first and avoids the more-processor-intensive (ChessBoard.move())
logic until it is absolutely necessary to call.